updates validate script for new build paths
